Move all treemap viewer specific code to reporter-treemap directory
This is the first step in separating the reporter from the webpack
specific parts
