+++ reporter-treemap/client/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../.eslintrc.json",
+  "extends": "../../.eslintrc.json",
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
