+++ reporter-treemap/client/components/ModulesTreemap.jsx
