+++ reporter-treemap/client/components/Sidebar.css
