+++ reporter-treemap/client/components/Sidebar.jsx
