+++ reporter-treemap/client/components/Switcher.css
