+++ reporter-treemap/client/components/Switcher.jsx
