+++ reporter-treemap/client/components/SwitcherItem.jsx
