+++ reporter-treemap/client/viewer.jsx
