+++ src/bin/analyzer.js
@@ -8,7 +8,8 @@ const commander = require('commander');
 const { magenta } = require('chalk');
 
 const getChartData = require('../chartData');
-const viewer = require('../viewer');
+// TODO: Inject this module somehow
+const viewer = require('../../reporter-treemap');
 const Logger = require('../Logger');
 
 const program = commander
